+++ components/pages/pub/faq/Faq.data.ts
@@ -1,1 +1,0 @@
-export * from '../../faq/Faq.data';