+++ src/app/pub/admin/faqs/categories/page.tsx
@@ -1,7 +1,0 @@
-import FaqCategories from '../../../../../../components/pages/pub/admin/FaqCategories';
-
-const Page = () => {
-  return <FaqCategories />;
-};
-
-export default Page;