+++ src/app/pub/admin/faqs/page.tsx
@@ -1,7 +1,0 @@
-import FaqList from '../../../../../components/pages/pub/admin/Admin';
-
-const Page = () => {
-  return <FaqList />;
-};
-
-export default Page;