+++ src/app/pub/admin/faqs/register/page.tsx
@@ -1,7 +1,0 @@
-import FaqRegister from '../../../../../../components/pages/pub/admin/FaqRegister';
-
-const Page = () => {
-  return <FaqRegister />;
-};
-
-export default Page;