+++ src/app/pub/admin/login/page.tsx
@@ -1,7 +1,0 @@
-import AdminLogin from '../../../../../components/pages/pub/admin/login/AdminLogin';
-
-const Page = () => {
-  return <AdminLogin />;
-};
-
-export default Page;