+++ src/app/pub/admin/page.tsx
@@ -1,7 +1,0 @@
-import Admin from '../../../../components/pages/pub/admin/Admin';
-
-const Page = () => {
-  return <Admin />;
-};
-
-export default Page;